Move connect code from main activity to fragment Chang ui remote control
