+++ app/src/main/res/drawable/border_rectangle.xml
@@ -1,0 +1,4 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle" >
+    <solid android:color="@android:color/white" />
+    <stroke android:width="5dp" android:color="@color/colorPrimary"/>
+</shape>
