+++ app/src/main/res/drawable/bt_power.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" xmlns:aapt="http://schemas.android.com/aapt"
+    android:viewportWidth="24"
+    android:viewportHeight="24"
+    android:width="24dp"
+    android:height="24dp">
+    <path
+        android:pathData="M13 3h-2v10h2V3zm4.83 2.17l-1.42 1.42C17.99 7.86 19 9.81 19 12c0 3.87 -3.13 7 -7 7s-7 -3.13 -7 -7c0 -2.19 1.01 -4.14 2.58 -5.42L6.17 5.17C4.23 6.82 3 9.26 3 12c0 4.97 4.03 9 9 9s9 -4.03 9 -9c0 -2.74 -1.23 -5.18 -3.17 -6.83z"
+        android:fillColor="#000000" />
+</vector>
