+++ app/src/main/res/drawable/bt_rhythm.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="1280.000000dp"
+    android:height="640.000000dp"
+    android:viewportWidth="1280.000000"
+    android:viewportHeight="640.000000">
+
+    <group
+        android:translateY="640.000000"
+        android:scaleX="0.100000"
+        android:scaleY="-0.100000">
+        <path
+            android:fillColor="#000000"
+            android:strokeWidth="1"
+            android:pathData="M6465 6336 c-13 -13 -26 -35 -29 -48 -2 -12 -14 -147 -25 -298 -19 -251 -64 -822 -156 -1995 -19 -242 -50 -640 -70 -885 -19 -245 -57 -728 -84 -1075 -28 -346 -58 -736 -67 -865 -10 -129 -21 -239 -26 -245 -4 -5 -8 2 -8 15 0 14 -45 324 -100 690 -55 366 -143 951 -196 1300 -52 349 -112 749 -134 888 -40 265 -46 284 -96 296 -13 4 -39 0 -58 -8 -42 -17 -30 20 -167 -514 l-104 -402 -400 0 c-361 0 -883 4 -3807 26 l-877 6 -31 -30 c-38 -38 -38 -72 -2 -112 l27 -31 1390 -10 c765 -5 1925 -13 2580 -18 1272 -9 1241 -10 1260 37 4 9 38 134 75 277 56 218 69 254 74 225 7 -40 43 -278 115 -760 28 -184 105 -695 171 -1135 66 -440 145 -966 176 -1170 30 -203 57 -378 60 -388 17 -57 96 -76 139 -33 29 29 30 37 50 306 13 173 94 1215 145 1860 5 69 37 469 70 890 64 812 76 962 115 1460 14 173 30 386 37 472 6 87 13 156 15 154 2 -2 60 -620 128 -1374 69 -754 129 -1381 134 -1394 13 -36 39 -52 78 -51 61 2 98 52 251 343 l141 265 2738 5 c2582 5 2740 6 2765 22 35 23 47 78 24 113 -9 13 -33 29 -53 35 -27 8 -828 10 -2795 8 l-2758 -3 -24 -25 c-14 -14 -70 -110 -125 -215 -55 -104 -101 -188 -102 -185 0 3 -74 803 -163 1779 l-162 1774 -28 23 c-36 31 -80 31 -111 0z" />
+    </group>
+</vector>
