+++ app/src/main/res/drawable/bt_swing.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="1000dp"
+    android:height="1000dp"
+    android:viewportWidth="1000"
+    android:viewportHeight="1000">
+
+    <path
+        android:fillColor="#000000"
+        android:pathData="M959.6,269.3H257.8V151.2c0-5.6-3.2-10.9-8.2-13.4c-5-2.7-11-2.4-15.7,0.9L16.7,287.2c-4.1,2.8-6.7,7.6-6.7,12.6c0,5,2.6,9.8,6.6,12.6L233.8,461c2.6,1.8,5.6,2.7,8.7,2.7c2.4,0,4.9-0.6,7.1-1.8c5-2.6,8.2-7.8,8.2-13.4V330.3h701.8c17,0,30.5-13.7,30.5-30.5C990,283,976.5,269.3,959.6,269.3z" />
+    <path
+        android:fillColor="#000000"
+        android:pathData="M983.4,687.4L766.2,538.7c-2.6-1.7-5.5-2.6-8.5-2.6c-2.6,0-4.8,0.6-7.2,1.8c-5,2.6-8,7.7-8,13.4v118.2H40.5c-17,0-30.5,13.5-30.5,30.5c0,17.1,13.5,30.5,30.5,30.5h701.8v118.2c0,5.9,3,11,7.9,13.4c5,2.7,11.1,2.4,15.9-0.9l217.1-148.7c4.1-2.8,6.8-7.7,6.8-12.7C990,695,987.6,690.3,983.4,687.4z" />
+</vector>
