+++ app/src/main/res/drawable/bt_white_round.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android = "http://schemas.android.com/apk/res/android">
+    <item android:state_pressed = "false">
+        <shape android:shape = "oval">
+            <solid android:color = "#FFFFFF"/>
+        </shape>
+    </item>
+</selector>
