+++ app/src/main/res/layout/fragment_remote_control.xml
@@ -1,48 +1,112 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:orientation="vertical"
+    android:layout_width="match_parent"
     android:layout_height="match_parent">
 
+
     <TextView
-        android:id="@+id/text_remote_control"
+        android:id="@+id/tv_remote_control"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textAlignment="center"
-        android:textSize="22sp"
-        android:layout_margin="8dp" />
+        android:layout_height="40dp"
+        android:layout_marginLeft="60dp"
+        android:layout_marginRight="60dp"
+        android:layout_marginTop="20dp"
+        android:gravity="center"
+        android:text="Test"
+        android:textSize="25sp"
+        android:background="@drawable/border_rectangle"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+
     <Button
         android:id="@+id/bt_power"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:onClick="clickFanRemoteButton"
-        android:text="@string/fan_power" />
+        android:layout_width="150sp"
+        android:layout_height="150sp"
+        android:layout_marginTop="350dp"
+        android:background="@drawable/bt_white_round"
+        android:foreground="@drawable/bt_power"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <Button
         android:id="@+id/bt_swing"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:onClick="clickFanRemoteButton"
-        android:text="@string/fan_swing" />
+        android:layout_width="100sp"
+        android:layout_height="100sp"
+        android:layout_marginTop="240dp"
+        android:background="@drawable/bt_white_round"
+        android:foreground="@drawable/bt_swing"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.18"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <Button
         android:id="@+id/bt_speed"
-        android:onClick="clickFanRemoteButton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/fan_speed" />
+        android:layout_width="100sp"
+        android:layout_height="100sp"
+        android:layout_marginTop="240dp"
+        android:background="@drawable/bt_white_round"
+        android:foreground="@drawable/bt_speed"
+        app:layout_constraintHorizontal_bias="0.842"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <Button
         android:id="@+id/bt_timer"
-        android:onClick="clickFanRemoteButton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/fan_timer" />
+        android:layout_width="100sp"
+        android:layout_height="100sp"
+        android:layout_marginTop="510dp"
+        android:background="@drawable/bt_white_round"
+        android:foreground="@drawable/bt_timer"
+        app:layout_constraintHorizontal_bias="0.18"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <Button
         android:id="@+id/bt_rhythm"
-        android:onClick="clickFanRemoteButton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/fan_rhythm" />
+        android:layout_width="100sp"
+        android:layout_height="100sp"
+        android:layout_marginTop="510dp"
+        android:foreground="@drawable/bt_rhythm"
+        android:background="@drawable/bt_white_round"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.82"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
-</LinearLayout>
+    <TextView
+        android:id="@+id/tv_status"
+        android:layout_width="match_parent"
+        android:layout_height="150dp"
+        android:layout_marginTop="80dp"
+        android:layout_marginLeft="40dp"
+        android:layout_marginRight="40dp"
+        android:padding="10dp"
+        android:textSize="20sp"
+        android:background="@drawable/border_rectangle"
+        app:layout_constraintHorizontal_bias="0.15"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
